Move MockIOTile device into IOTile-Core and rename
- Also add virtualinterface to iotilecore
