+++ iotiletest/tox.ini
@@ -5,4 +5,5 @@ envlist = py
 passenv=APPDATA
 deps=
     pytest
+    ../iotilecore
 commands=py.test
